analyzer: move addresses into util/addresses
